Created urls file for the Accounts App
