+++ .env.sample
@@ -1,7 +1,7 @@
 # Configuration Mode => development, testing, staging, or production
 CONFIG_MODE = development
 
-# SQLALCHEMY_DATABASE_URI => 'postgresql+psycopg2://user:password@host/database'
+#  => 'postgresql+psycopg2://user:password@host/database'
 DEVELOPMENT_DATABASE_URL = 'postgresql+psycopg2://testuser:testpass@localhost:5432/flask_sqlalchemy_postgresql'
 TEST_DATABASE_URL        =
 STAGING_DATABASE_URL     =
